api conversion for Airsim 1.3.1
