+++ gym_airsim/__init__.py
@@ -1,7 +1,5 @@
 from gym.envs.registration import register
-"""
 register(
     id='AirSimEnv-v42',
     entry_point='gym_airsim.envs:AirSimEnv',
 )
-"""
