+++ gym_airsim/envs/AirGym.py
@@ -10,7 +10,6 @@ from gym.spaces.box import Box
 
 from gym_airsim.envs.myAirSimClient import *
         
-from AirSimClient import *
 
 logger = logging.getLogger(__name__)
